requirements.txt
fastapi==0.102.0
uvicorn[standard]==0.23.1
starlette==0.27.0
pydantic==1.10.11
langgraph
tavily-python==0.1.0
openai>=0.27.0,<0.30.0
python-dotenv==1.0.0
pymongo==4.3.3
requests==2.31.0
